Transmition speed Improved WriteNoResponse
Now it sends data to server in Write Without Response mode. This increase the transmission speed and it get adapted to MIDI over BLE protocol.
Some devices needs a first message with confirmation for clean security flags. After a connection or after an error, the first message is sended like Write (with response).