Change begin method to inline in BLEMIDI_ESP32 to prevent linker errors
Add inline to begin() to prevent linker errors

When using the `lathoub/BLE-MIDI` library in a PlatformIO project with `ESPAsyncWebServer`, a "multiple definition" linker error occurs for the `BLEMIDI_ESP32::begin` function. 
This is because the function is defined in the header file `hardware/BLEMIDI_ESP32.h`, and this header is included in multiple source files.